user MS, MrMs exception handling

+++ MedicalRecordMicroservice/src/main/java/com/medimate/MedicalRecordMicroservice/services/MedicalRecordService.java
@@ -2,19 +2,22 @@ package com.medimate.MedicalRecordMicroservice.services;
 
 import com.medimate.MedicalRecordMicroservice.models.MedicalRecord;
 import com.medimate.MedicalRecordMicroservice.repositories.MedicalRecordRepository;
+import com.medimate.MedicalRecordMicroservice.repositories.PatientRepository;
 import com.medimate.MedicalRecordMicroservice.viewmodels.MedicalRecordVM;
 import jakarta.persistence.EntityExistsException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 
 @Service
 public class MedicalRecordService {
 
     @Autowired
     private MedicalRecordRepository medicalRecordRepository;
+
+    @Autowired
+    private PatientRepository patientRepository;
 
     public MedicalRecord addMedicalRecord(MedicalRecordVM medicalRecordVM)
     {
@@ -51,16 +54,30 @@ public class MedicalRecordService {
 
     public void deleteMedicalRecordForPatient(Integer id, Integer patientId)
     {
+        if(patientRepository.findById(patientId).orElse(null) == null) {
+            throw new EntityExistsException("Could not find patient with id " + patientId);
+        }
+
+        if(medicalRecordRepository.findById(id).orElse(null) == null){
+            throw new EntityExistsException("Could not find medical record with id " + patientId);
+        }
+
         medicalRecordRepository.deleteByPatientId(id, patientId);
     }
 
     public void deleteAllMedicalRecordsForPatient(Integer patientId)
     {
+        if(patientRepository.findById(patientId).orElse(null) == null){
+            throw new EntityExistsException("Could not find patient with id " + patientId);
+        }
         medicalRecordRepository.deleteAllByPatientId(patientId);
     }
 
     public void deleteMedicalRecord(Integer id)
     {
+        if(medicalRecordRepository.findById(id).orElse(null) == null){
+            throw new EntityExistsException("Could not find medical record with id " + id);
+        }
             medicalRecordRepository.deleteById(id);
     }
 }
